refactoring. transfer /mint to /
